Moving Entity and Entities to common.
